// apps/enterprise-service-management/src/generated/models/service-category-model.ts
import type { Workspace } from './workspace-model';

export const ServiceCategoryStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type ServiceCategoryStatusKey = keyof typeof ServiceCategoryStatusKeyToLabel;

export interface ServiceCategory {
  /**
   * @displayName Service Category ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Icon
   * @validationRule Required for create/update operations
   */
  icon: string;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ServiceCategoryStatusKey;
  /**
   * @displayName Updated At
   */
  updatedAt?: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _ServiceCategory = 'ServiceCategory' as const;